make prometheus client more modular
